requirements.txt
chromadb
openai
loguru
pydantic
tenacity
llama-index
fastapi
uvicorn
python-dotenv